Refactor UIManager to delegate tooltip and analysis panel management to TooltipManager and AnalysisPanelManager; implement AIPipelineConfig and AIPipelineLoader for AI pipeline management; create BackgroundContextMenu and BackgroundMessageHandler for context menu interactions and message processing; develop PopupAnalyzer and PopupResultRenderer for popup analysis operations; enhance entity highlighting with EntityHighlighter; establish TooltipManager for tooltip display; overall code organization and modularization for improved maintainability and readability.
