+++ examples/kitchensink-extension/src/ai-manager.ts
@@ -1,22 +1,18 @@
-// ai-manager.ts
-import { pipeline, Pipeline, env } from '@xenova/transformers';
+/**
+ * AIManager - Main coordinator for AI operations
+ */
+
+import { env } from '@xenova/transformers';
 import type {
-  PipelineConfig,
-  PipelineResult,
-  ExtensionMessage,
-  ExtensionResponse,
-  MessageHandler,
-  PageData,
-  ChatMessage,
-  AIManagerState,
   TaskType,
   SentimentResult,
   NamedEntity,
   SummarizationResult,
   ClassificationResult,
-  AnalysisResult,
-  QuestionAnsweringResult
+  AIManagerState
 } from './types';
+import { AIPipelineConfig } from './ai-pipeline-config';
+import { AIPipelineLoader } from './ai-pipeline-loader';
 
 // Configure Transformers.js for extension environment
 env.allowRemoteModels = false;
@@ -25,7 +21,8 @@ env.backends.onnx.wasm.wasmPaths = chrome.runtime.getURL('dist/');
 
 export class AIManager {
   private state: AIManagerState;
-  private readonly configs: Map<TaskType, PipelineConfig>;
+  private config: AIPipelineConfig;
+  private loader: AIPipelineLoader;
 
   constructor() {
     this.state = {
@@ -35,155 +32,77 @@ export class AIManager {
       isInitialized: false
     };
 
-    // Define pipeline configurations
-    this.configs = new Map([
-      ['sentiment', {
-        name: 'sentiment-analysis',
-        task: 'sentiment-analysis',
-        model: 'Xenova/distilbert-base-uncased-finetuned-sst-2-english',
-        quantized: true,
-        status: 'loading'
-      }],
-      ['ner', {
-        name: 'token-classification',
-        task: 'token-classification',
-        model: 'Xenova/distilbert-base-NER',
-        quantized: true,
-        status: 'loading'
-      }],
-      ['summarization', {
-        name: 'summarization',
-        task: 'summarization',
-        model: 'Xenova/distilbart-cnn-6-6',
-        quantized: true,
-        status: 'loading'
-      }],
-      ['classification', {
-        name: 'text-classification',
-        task: 'text-classification',
-        model: 'Xenova/distilbert-base-uncased-finetuned-sst-2-english',
-        quantized: true,
-        status: 'loading'
-      }],
-      ['qa', {
-        name: 'question-answering',
-        task: 'question-answering',
-        model: 'Xenova/distilbert-base-cased-distilled-squad',
-        quantized: true,
-        status: 'loading'
-      }],
-      ['embedding', {
-        name: 'feature-extraction',
-        task: 'feature-extraction',
-        model: 'Xenova/all-MiniLM-L6-v2',
-        quantized: true,
-        status: 'loading'
-      }]
-    ]);
-
+    this.config = new AIPipelineConfig();
+    this.loader = new AIPipelineLoader(this.state, this.config);
     this.init();
   }
 
   private async init(): Promise<void> {
     console.log('🤖 AI Context Copilot: Initializing AI Manager...');
     try {
-      await this.loadPipeline('sentiment');
-      await this.loadPipeline('ner');
+      await this.loader.loadCoreModels();
       this.state.isInitialized = true;
       console.log('✅ AI Manager initialized successfully');
-      this.loadRemainingPipelines();
+      this.loader.loadSecondaryModels();
     } catch (error) {
       console.error('❌ Failed to initialize AI Manager:', error);
     }
   }
 
-  private async loadRemainingPipelines(): Promise<void> {
-    const remaining: TaskType[] = ['summarization', 'classification', 'qa', 'embedding'];
-    for (const taskType of remaining) {
-      try {
-        await this.loadPipeline(taskType);
-        await new Promise(resolve => setTimeout(resolve, 1000));
-      } catch (error) {
-        console.warn(`Failed to load ${taskType} pipeline:`, error);
-      }
-    }
-  }
-
-  private async loadPipeline(taskType: TaskType): Promise<any> {
-    const config = this.configs.get(taskType);
-    if (!config) throw new Error(`No configuration found for task: ${taskType}`);
-    const existingPipeline = this.state.pipelines.get(taskType);
-    if (existingPipeline) return existingPipeline;
-    console.log(`📥 Loading ${taskType} pipeline...`);
-    this.state.status.set(taskType, 'loading');
-    try {
-      const pipelineInstance = await pipeline(
-        config.task,
-        config.model,
-        { quantized: config.quantized }
-      );
-      this.state.pipelines.set(taskType, pipelineInstance);
-      this.state.status.set(taskType, 'loaded');
-      config.status = 'loaded';
-      config.lastUsed = Date.now();
-      console.log(`✅ ${taskType} pipeline loaded successfully`);
-      return pipelineInstance;
-    } catch (error) {
-      console.error(`❌ Failed to load ${taskType} pipeline:`, error);
-      this.state.status.set(taskType, 'error');
-      config.status = 'error';
-      throw error;
-    }
-  }
-
   async analyzeSentiment(text: string): Promise<SentimentResult[]> {
-    const pipelineInstance = await this.loadPipeline('sentiment');
-    const result = await pipelineInstance(text);
+    const pipeline = await this.loader.loadPipeline('sentiment');
+    const result = await pipeline(text);
     return Array.isArray(result) ? result : [result];
   }
 
   async extractEntities(text: string): Promise<NamedEntity[]> {
-    const pipelineInstance = await this.loadPipeline('ner');
-    const result = await pipelineInstance(text);
+    const pipeline = await this.loader.loadPipeline('ner');
+    const result = await pipeline(text);
     return Array.isArray(result) ? result : [result];
   }
 
   async summarizeText(text: string): Promise<SummarizationResult> {
-    const pipelineInstance = await this.loadPipeline('summarization');
+    const pipeline = await this.loader.loadPipeline('summarization');
     const maxLength = 1024;
     const inputText = text.length > maxLength ? text.substring(0, maxLength) : text;
-    const result = await pipelineInstance(inputText, {
+    
+    const result = await pipeline(inputText, {
       max_length: 150,
       min_length: 30,
       do_sample: false
     });
+    
     return Array.isArray(result) ? result[0] : result;
   }
 
   async classifyText(text: string): Promise<ClassificationResult[]> {
-    const pipelineInstance = await this.loadPipeline('classification');
-    const result = await pipelineInstance(text);
+    const pipeline = await this.loader.loadPipeline('classification');
+    const result = await pipeline(text);
     return Array.isArray(result) ? result : [result];
   }
 
   async answerQuestion(question: string, context: string): Promise<any> {
-    const pipelineInstance = await this.loadPipeline('qa');
-    return await pipelineInstance(question, context);
+    const pipeline = await this.loader.loadPipeline('qa');
+    return await pipeline(question, context);
   }
 
   async generateEmbedding(text: string): Promise<number[]> {
-    const pipelineInstance = await this.loadPipeline('embedding');
-    const result = await pipelineInstance(text, { pooling: 'mean', normalize: true });
+    const pipeline = await this.loader.loadPipeline('embedding');
+    const result = await pipeline(text, { pooling: 'mean', normalize: true });
     return result.data;
   }
 
   async analyzeText(text: string, taskType: TaskType): Promise<any> {
-    if (!text || text.trim().length === 0) throw new Error('No text provided for analysis');
+    if (!text || text.trim().length === 0) {
+      throw new Error('No text provided for analysis');
+    }
+
     const cacheKey = `${taskType}:${text.substring(0, 100)}`;
     const cached = this.state.cache.get(cacheKey);
     if (cached && Date.now() - cached.timestamp < 300000) {
       return cached.data;
     }
+
     let result: any;
     switch (taskType) {
       case 'sentiment':
@@ -207,19 +126,17 @@ export class AIManager {
       default:
         throw new Error(`Unsupported task type: ${taskType}`);
     }
+
     this.state.cache.set(cacheKey, {
       data: result,
       timestamp: Date.now()
     });
+
     return result;
   }
 
   getStatus(): Record<string, string> {
-    const status: Record<string, string> = {};
-    for (const [taskType, pipelineStatus] of this.state.status) {
-      status[taskType] = pipelineStatus;
-    }
-    return status;
+    return this.loader.getStatus();
   }
 
   isReady(): boolean {
